+++ voLAWNteer/Controllers/LawnsController.cs
@@ -43,6 +43,28 @@ namespace voLAWNteer.Controllers
             return View(lawn);
         }
 
+        // GET: Lawns/Create
+        public IActionResult Create()
+        {
+            return View();
+        }
+
+        // POST: Lawns/Create
+        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
+        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Create([Bind("Id,FirstName,LastName,StreetAddress,City,State,ZipCode,Size,Description,Approved,Photo")] Lawn lawn)
+        {
+            ModelState.Remove("lawn.approved");
+            if (ModelState.IsValid)
+            {
+                _context.Add(lawn);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            return View(lawn);
+        }
 
         // GET: Lawns/Edit/5
         public async Task<IActionResult> Edit(int? id)
