+++ voLAWNteer/Models/Lawn.cs
@@ -41,8 +41,7 @@ namespace voLAWNteer.Models
         [Required]
         public string Description { get; set; }
         
-        [Required]
-        public bool Approved { get; set; }
+        public bool? Approved { get; set; }
         
         [Required]
         public string Photo { get; set; }
